+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Properly implement definition through LSP API.
+
 ## [v3.14.3](https://github.com/obsidian-nvim/obsidian.nvim/releases/tag/v3.14.3) - 2025-10-19
 
 ### Fixed